+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/state/global/internal/MutableGlobalState.kt
@@ -20,8 +20,7 @@ import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.models.ChannelMute
 import io.getstream.chat.android.client.models.Mute
 import io.getstream.chat.android.client.models.TypingEvent
-import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.offline.model.connection.ConnectionState
+import io.getstream.chat.android.client.setup.state.ClientState
 import io.getstream.chat.android.offline.plugin.state.global.GlobalState
 import io.getstream.chat.android.offline.utils.Event
 import io.getstream.chat.android.state.plugin.internal.StatePlugin
@@ -31,13 +30,9 @@ import io.getstream.chat.android.state.plugin.internal.StatePlugin
  */
 internal interface MutableGlobalState : GlobalState {
 
+    val clientState: ClientState
+
     fun setErrorEvent(errorEvent: Event<ChatError>)
-
-    fun setUser(user: User)
-
-    fun setConnectionState(connectionState: ConnectionState)
-
-    fun setInitialized(initialized: Boolean)
 
     fun setTotalUnreadCount(totalUnreadCount: Int)