+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/feature/chat/info/shared/media/ChatInfoSharedMediaFragment.kt
@@ -27,7 +27,6 @@ import androidx.lifecycle.Transformations
 import androidx.lifecycle.asLiveData
 import androidx.navigation.fragment.navArgs
 import io.getstream.chat.android.client.ChatClient
-import io.getstream.chat.android.offline.extensions.globalState
 import io.getstream.chat.android.ui.ChatUI
 import io.getstream.chat.android.ui.gallery.AttachmentGalleryDestination
 import io.getstream.chat.android.ui.gallery.AttachmentGalleryItem
@@ -110,7 +109,7 @@ class ChatInfoSharedMediaFragment : Fragment() {
 
     private fun bindViewModel() {
         Transformations.switchMap(viewModel.state) { state ->
-            Transformations.map(ChatClient.instance().globalState.user.asLiveData()) { user ->
+            Transformations.map(ChatClient.instance().clientState.user.asLiveData()) { user ->
                 user to state
             }
         }.observe(viewLifecycleOwner) { (user, state) ->